Renaming functions, reorganizing code, adding temporary hack for closing gdbserver
